starpu: silent warnings

+++ runtime/starpu/interface/cham_tile_interface.c
@@ -119,7 +119,7 @@ cti_register_data_handle( starpu_data_handle_t  handle,
                           void                 *data_interface )
 {
     starpu_cham_tile_interface_t *cham_tile_interface = (starpu_cham_tile_interface_t *) data_interface;
-    unsigned node;
+    int node;
 
     for (node = 0; node < STARPU_MAXNODES; node++)
     {
@@ -573,10 +573,10 @@ static int cti_copy_any_to_any( void *src_interface, unsigned src_node,
     starpu_cham_tile_interface_t *cham_tile_src = (starpu_cham_tile_interface_t *) src_interface;
     starpu_cham_tile_interface_t *cham_tile_dst = (starpu_cham_tile_interface_t *) dst_interface;
     size_t elemsize = CHAMELEON_Element_Size( cham_tile_src->flttype );
-    size_t m        = cham_tile_src->tile.m;
-    size_t n        = cham_tile_src->tile.n;
-    size_t ld_src   = cham_tile_src->tile.ld;
-    size_t ld_dst   = cham_tile_dst->tile.ld;
+    size_t m        = (size_t)(cham_tile_src->tile.m);
+    size_t n        = (size_t)(cham_tile_src->tile.n);
+    size_t ld_src   = (size_t)(cham_tile_src->tile.ld);
+    size_t ld_dst   = (size_t)(cham_tile_dst->tile.ld);
     int    ret      = 0;
 
     void *src_mat = CHAM_tile_get_ptr( &(cham_tile_src->tile) );
@@ -585,8 +585,8 @@ static int cti_copy_any_to_any( void *src_interface, unsigned src_node,
     assert( ld_src >= m );
     assert( ld_dst >= m );
 
-    assert( m == cham_tile_dst->tile.m );
-    assert( n == cham_tile_dst->tile.n );
+    assert( m == (size_t)(cham_tile_dst->tile.m) );
+    assert( n == (size_t)(cham_tile_dst->tile.n) );
 
 #if defined(CHAMELEON_KERNELS_TRACE)
     fprintf( stderr,
